refactor: rename svg with props
