+++ src/components/svgs/eth-change-arrow-with-props/EthChangeArrowUp.js
