+++ src/components/svgs/eth-change-arrow-with-props/EthChangeArrowZero.js
