+++ src/components/svgs/spring-with-props/Spring0.js
