+++ src/components/svgs/spring-with-props/SpringMinus1.js
