+++ src/components/svgs/spring-with-props/SpringMinus10.js
