+++ src/components/svgs/spring-with-props/SpringMinus2.js
