+++ src/components/svgs/spring-with-props/SpringMinus3.js
