+++ src/components/svgs/spring-with-props/SpringMinus4.js
