+++ src/components/svgs/spring-with-props/SpringMinus5.js
