+++ src/components/svgs/spring-with-props/SpringMinus6.js
