+++ src/components/svgs/spring-with-props/SpringMinus7.js
