+++ src/components/svgs/spring-with-props/SpringMinus8.js
