+++ src/components/svgs/spring-with-props/SpringMinus9.js
