+++ src/components/svgs/spring-with-props/SpringPlus1.js
