+++ src/components/svgs/spring-with-props/SpringPlus10.js
