+++ src/components/svgs/spring-with-props/SpringPlus2.js
