+++ src/components/svgs/spring-with-props/SpringPlus3.js
