+++ src/components/svgs/spring-with-props/SpringPlus4.js
