+++ src/components/svgs/spring-with-props/SpringPlus5.js
