+++ src/components/svgs/spring-with-props/SpringPlus6.js
