+++ src/components/svgs/spring-with-props/SpringPlus7.js
