+++ src/components/svgs/spring-with-props/SpringPlus8.js
