+++ src/components/svgs/spring-with-props/SpringPlus9.js
