+++ src/components/svgs/spring-with-props/index.js
