+++ src/components/svgs/thermometer-with-props/Thermometer.js
