+++ src/components/svgs/thermometer-with-props/index.js
